requirements.txt
coverage
flask
requests
numpy
pandas
matplotlib
mysql-connector-python
